text2sql/backend/requirements.txt
fastapi
pydantic
uvicorn
langchain
langchain_community
ibm-watson
ibm-watson-machine-learning
pandas
urllib3<2
python-dotenv
faiss-cpu
huggingface_hub
openpyxl
python-multipart
python-dotenv